Add support to different image types
A new type (DTImageType) was created to represent the image format,
together with a new routine to identify the format according to a file
header (IdentifyImageType).

These modifications will facilitate future maintenance and addition of
even more supported formats.

This commit also adds the `png.h` header, and uses the library call to
check the PNG header.